Use modified settings file for integration tests
During integration tests we use remote repositories
inside the local file system. These repos are mirrored
by the mirror usually defind in the settings file.
In order to avoid this we copy the settings file, and change
the mirrorOf string in order not to include local repositories.
Afterwards the modified settings file is used by the integation
tests.