Support for non default settings files
Up to now the integration tests worked only upon the
settings file on the default location ~/.m2/settings.

Now it is possible to provide alternate settings files.
The alternate settings file needs to be provided with -s flag
to maven and again via property "com.sap.maven.integration-tests.user-settings".

Also the local repository is shared between thatone used for the build of the
xcode-maven-plugin and thatone used during the integration tests. This is
important since otherwise the xcode-maven-plugin could either not be found at
all or a wrong version might be found.