Keep the build settings in memory
Up to now the build settings was extracted by a dedicated
mojo. Now the are extraced by the EffectiveBuildSettings class
directly. The settings are kept in memory.

In order to achive that the AbstracXCodeBuildMojo has been renamed
to BuildContextAwareMojo in order to have better naming.

All classes that requires access to build settings are now BuildContextAwareMojos.