Provide unarchiverId from outside.
With the current approach regarding handling of additional bundles/archives
we might run into problems in case of a not commonly used file appendix.
Up to now we use the file appendix of the main artifact in order to derive the
unarchiver.

With the approach implemented here is it possible to provide the unarchiver from
outside. In case no unarchiver is provided the unarchiver is still derived from the
file appendix of the main artifact.