Provide mojo for validation checks for xcode applications
With this mojo xcode applications build with the plugin in the compile
phase can be validated (e.g. sizes of the images used by the plugin).

The checks itself are not part of the xcode-maven-plugin. The checks needs
to be provided in one or more separate jar file(s). Checks are configured
in an xml file. Documentation will be provided with a later commit.

is already provided with the xcode context.